refactor db, add alembic. Add sqlite to poetry
