+++ pyproject.toml
@@ -18,6 +18,7 @@ python-dotenv = "^1.0.1"
 pydantic = {extras = ["email"], version = "^2.9.0"}
 sqlalchemy = "^2.0.34"
 alembic = "^1.13.2"
+aiosqlite = "^0.20.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.6.4"
