+++ src/db/database.py
@@ -1,13 +1,10 @@
 from typing import AsyncGenerator
 
-from sqlalchemy import MetaData
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 from sqlalchemy.ext.declarative import declarative_base
 
-from src.settings import settings
+from config import settings
 
-
-metadata = MetaData()
 
 engine = create_async_engine(settings.DATABASE_URL, echo=True, future=True)
